Move types to utilities file.
